Cards.jl Documentation

Overview

This module describes the enums/structs/functions to play a simple version of Poker.

Enums:

Suit – The 4 suits of the cards.

Rank – The 13 different ranks.

PokerType – The different poker hands one may have.

Structs:

Card – A description of a given card as a combination of Suit and Rank.

PokerHand – A data structure describing a poker hand including its classification; i.e., its PokerType.

Deck – A description of a deck of cards.

Functions:

Deck Functions:

Deck() – Create a deck of cards.

shuffle_deck! – Shuffle the deck of cards.

restore_deck! – Return all "dealt" cards back to the place in the deck where they were drawn.

num_cards_left_in_deck – The number of cards left in the deck.

Card Play

draw_cards! – Draw some number of cards from a deck.

deal_hand! – Deal a poker hand to a player.

make_secondary_draw! – Execute a second draw of a simple poker game.

play_poker! – Simulate a simple version of a poker game with two players.

There is an associated Jupyter notebook at src/CardTest.ipynb.

Enums

Cards.Suit — Type

Suit (Enum)

Allowable suits in poker; ordered by strength from lowest to highest.

Fields

♠

♦

♣

♥

Cards.Rank — Type

Rank (Enum)

Card ranks in poker; ordered by strength from lowest to highest.

Fields

Two Three Four Five Six Seven Eight Nine Ten

Jack

Queen

King

Ace

Cards.PokerType — Type

PokerType (Enum)

Enumeration describing a classification of hands for poker; ordered by strength from lowest to highest.

Fields

HighCard

OnePair

TwoPair

ThreeOfAKind

Straight

Flush

FullHouse

FourOfAKind

StraightFlush

Structs

Cards.Card — Type

Card

DATA STRUCTURE: Representation of a Card.

Fields

suit :: Suit

rank :: Rank

Constructors

Card(::Suit, ::Rank)

Cards.PokerHand — Type

PokerHand

DATA STRUCTURE: Representation of a Card.

This inner constructor takes a vector of cards:

Checks the input contract:

Checks that the number of cards is correct;

Checks that there are no duplicates.

Sorts the cards (isless is defined for Card).

Creates an internal representation grouping cards by Rank.

Classifies the hand into one of the poker types: PokerType. These are the familiar names of poker hands: TwoPair, FullHouse, etc.

Fields

cards :: Vector{Card}

gr_rep :: Vector{Tuple{Int, Rank}}

class :: PokerType

Constructor

PokerHand(::Vector{Card}; N::Int=5)

Cards.Deck — Type

Deck

MUTABLE DATA STRUCTURE: Representation of a deck of cards.

Fields

place :: Int – The place in the deck beyond which we may draw.

cds :: Vector{Card} – A vector of cards.

Constructors

Deck() – Creates the standard 52 card deck using the standard Card ordering.

Deck(place::Int, cards::Vector{Card}) – Create a deck of cards manually.

Functions

Cards.shuffle_deck! — Function

shuffle_deck!(d)

Shuffles the Deck, d, destructively; that is, the deck is changed as a result of this function.

This function only shuffles "what's left" of the deck. It will not interfere with the function restore_deck!, in the sense that all of the cards will put back, but the "restore" does not interfere with current and previous shuffle_deck!.

Arguments

d :: Deck – The deck to shuffle.

Return

nothing

Cards.restore_deck! — Function

restore_deck!(d)

Reset the Deck, d, to have all of the cards placed back into the deck.

Assumptions

The deck d has not been directly manipulated; that is, only the functions in this module should be used to manipulate a Deck.

Note

The cards that have been previous delt will be replaced; however,

previous calls to shuffle_deck! will remain in effect.

Arguments

d :: Deck – The Deck to operate on.

Return

::Nothing

Cards.num_cards_left_in_deck — Function

num_cards_left_in_deck(d)

Computes the number of cards left in deck, d.

Arguments

d :: Deck – The deck to query.

Return

The number of cards left in the deck.

Cards.deal_hand! — Function

deal_hand!(d; N=5)

Deals a hand from a deck, d, creating a PokerHand.

In the process, removes N cards from the deck, d.

Arguments

d :: Deck – A Deck from which to deal.

Keyword Arguments

N=5 :: Int – The number of cards to deal.

Return

::PokerHand – A poker hand

Cards.draw_cards! — Function

draw_cards!(d, N)

Draws N cards from a deck, d. In the process, removes N cards from the deck, d.

Arguments

d :: Deck – A Deck from which to draw.

N :: Int – The number of cards to draw.

Return

::Vector{Card} – A vector of Cards.

Cards.make_secondary_draw! — Function

make_secondary_draw!(h, d)

This function examines a poker hand, h, and decides (given that we can draw up to two cards from the deck) what cards (if any) to eliminate and then draw from the deck, d.

Although one may stick with the current hand, and consequently, not draw from the deck, we mark this function as mutable as it has the potential to mutate the deck, d.

Arguments

h :: PokerHand – The current poker hand to examine.

d :: Deck – The deck to work with.

Return

::PokerHand – A new poker hand. Potentially a copy of the original hand.

Cards.play_poker! — Function

play_poker!(d)

Play a game of poker with two players.

Process:

Deal each player two 5 card hands.

Compare two hands to see who would win.

Each player can then ask for up to 2 cards.

The two players are compared again to see who wins.

Note: This function mutates the deck by dealing cards to the players.

Arguments

d :: Deck – A deck of cards.

Return

::Nothing

Lower Level Functions

The Base function isless has been overloaded for the Structs: Card, and PokerHand.

This means that a vector of Card is sorted first by Suit and then Rank.

This means that a vector of PokerHand is sorted by the strength of the hand as measured by the game Poker.

The Base function (==) has been overloaded for struct PokerHand.

Cards.classify_hand — Function

classify_hand(gr_rep)

Classifies a poker hand into one of the standard classes given by the enumeration: PokerType.

This is done by first examining the length of the grouped rank representation, gr_rep.

We know the following:

|gr_rep| == 2 $\implies$ FourOfKind | FullHouse

|gr_rep| == 3 $\implies$ ThreeOfKind | TwoPair

|gr_rep| == 4 $\implies$ OnePair

|gr_rep| == 5 $\implies$ HighCard | Flush | Straight | StraightFlush

Arguments

gr_rep :: Vector{Tuple{Int, Rank}} – The internal representation of the hand. (See grouped_rank_rep).

cds :: Vector{Card} – Cards sorted by rank then suit.

Returns

::PokerType

Cards.grouped_rank_rep — Function

grouped_rank_rep(cds)

Creates a representation of a Poker hand.

The representation is a vector of two-tuples of the form: (N, Card-Rank) – Here, N is the number of times a card with Card-Rank appears in the hand. The tuple representation is ordered from highest to lowest.

NOTE: When N==1 the corresponding rank uniquely determines the card in the hand.

Input Contract

Cards are ASSUMED sorted via Base.isless(Card, Card).

Arguments

cds :: Vector{Card} – A Vector of Card.

Examples: The poker hand (♣ King, ♦ Jack, ♥ Nine, ♠ Nine, ♣ Three)

becomes: [(2, Nine), (1, King), (1, Jack), (1, Three)]

Return

::Vector{Tuple{Int, Rank}} – A Vector of two-tuples.

Cards.poker_hand_contract — Function

poker_hand_contract(cds; N=5)

Checks that a poker hand is valid.

Checks the following are true for cds:

They are unique.

Their number is N.

Arguments

cds :: Vector{Card} – A vector of cards.

Optional Arguments

N=5 :: Int – The number of cards that the hand should have.

Return

::Bool – true if cds are valid.

Index

Cards.Card

Cards.Deck

Cards.PokerHand

Cards.PokerType

Cards.Rank

Cards.Suit

Cards.classify_hand

Cards.deal_hand!

Cards.draw_cards!

Cards.grouped_rank_rep

Cards.make_secondary_draw!

Cards.num_cards_left_in_deck

Cards.play_poker!

Cards.poker_hand_contract

Cards.restore_deck!

Cards.shuffle_deck!